+++ android/src/main/kotlin/com/compress/all/flutter_compress/DownloadSaver.kt
@@ -60,4 +60,3 @@ object DownloadSaver {
         else -> "application/octet-stream"
     }
 }
-